Stop un-placing deleted signs.
There appeared to be no-benefit to un-placing signs that have been written over. And in certain cases, the act of un-placing each sign was dreadfully slow. This may be something to keep an eye on though, as a long running instance could quickly fill up with lots of dead signs.